LAB4/requirements.txt
PyOpenGL==3.1.7
pygame-ce==2.5.6
numpy==1.26.4
Pillow>=10.3.0